+++ posthog-react-native/test/storage.spec.ts
@@ -1,5 +1,5 @@
 import * as FileSystem from 'expo-file-system'
-import { SemiAsyncStorage } from '../src/storage'
+import { PostHogRNStorage } from '../src/storage'
 import { buildOptimisiticAsyncStorage } from '../src/native-deps'
 
 jest.mock('expo-file-system')
@@ -8,9 +8,8 @@ const mockedFileSystem = jest.mocked(FileSystem, true)
 describe('PostHog React Native', () => {
   jest.useRealTimers()
 
-  const storage = new SemiAsyncStorage(buildOptimisiticAsyncStorage())
-
   describe('storage', () => {
+    let storage: PostHogRNStorage
     beforeEach(() => {
       mockedFileSystem.readAsStringAsync.mockImplementation(() => {
         const res = Promise.resolve(
@@ -23,14 +22,13 @@ describe('PostHog React Native', () => {
         )
         return res
       })
+
+      storage = new PostHogRNStorage(buildOptimisiticAsyncStorage())
     })
 
     it('should load storage from the file system', async () => {
       expect(storage.getItem('foo')).toEqual(undefined)
-      storage.preloadAsync()
-      storage.preloadAsync()
-      storage.preloadAsync()
-      await storage.preloadAsync()
+      await storage.preloadPromise
       expect(mockedFileSystem.readAsStringAsync).toHaveBeenCalledTimes(1)
       expect(storage.getItem('foo')).toEqual('bar')
     })